change configuration copy process

+++ lib/supercop/actions/config_copier.rb
@@ -1,18 +1,19 @@
 module Supercop
   module Actions
-    class FileCreator
+    class ConfigCopier
       require 'fileutils'
+      CONFIG_PATH = %w[lib supercop templates].freeze
 
       def initialize(args)
         @filename = args.fetch(:filename, 'supercop.yml')
         @destination = args.fetch(:destination)
         @options = args.fetch(:options, {})
         @source = args.fetch(:source, 'supercop.yml')
-        @project = args.fetch(:project, Project)
       end
 
       def perform
         return "There is no destination #{destination}" if invalid_destination?
+        return "#{destination_file} already exists" if File.exist?(destination)
 
         FileUtils.copy(source_file, destination_file, options)
 
@@ -23,18 +24,22 @@ module Supercop
 
       private
 
-      attr_reader :filename, :destination, :source, :project, :options
+      attr_reader :filename, :destination, :source, :options
 
       def invalid_destination?
         destination.empty? || !Dir.exist?(destination)
       end
 
       def source_file
-        project.config_files(source)
+        File.join(root, CONFIG_PATH, source)
       end
 
       def destination_file
         File.join(destination, filename)
+      end
+
+      def root
+        Gem.loaded_specs['supercop'].full_gem_path
       end
     end
   end
